Move task code inside entry files
