+++ src/eval.py
@@ -1,15 +1,101 @@
-import hydra
 import pyrootutils
-from omegaconf import DictConfig
 
-root = pyrootutils.setup_root(__file__, dotenv=True, pythonpath=True)
+# -------------------------------------------------------------------------------------
+# `pyrootutils.setup_root(...)` is recommended at the top of each entry file!
+# https://github.com/ashleve/pyrootutils
+#
+# searches for ".git" or "pyproject.toml" in present and parent dirs
+# to determine the project root dir
+#
+# adds root dir to the PYTHONPATH (if `pythonpath=True`)
+# so this file can be run from any place without installing project as a package
+# otherwise modules would need to be imported relatively to the current work dir
+#
+# sets PROJECT_ROOT environment variable
+# used in "configs/paths/default.yaml"
+# this makes all paths relative to project root
+#
+# additionally loads environment variables from ".env" file (if `dotenv=True`)
+#
+root = pyrootutils.setup_root(
+    search_from=__file__,
+    indicator=[".git", "pyproject.toml"],
+    pythonpath=True,
+    dotenv=True,
+)
+#
+# you can get away without using `pyrootutils.setup_root(...)` if you:
+# - move this file to the project root dir
+# - always run this file from the project root dir
+# - modify paths in "configs/paths/default.yaml" to not use PROJECT_ROOT
+#
+# -------------------------------------------------------------------------------------
+
+from typing import List, Tuple
+
+import hydra
+from omegaconf import DictConfig
+from pytorch_lightning import LightningDataModule, LightningModule, Trainer
+from pytorch_lightning.loggers import LightningLoggerBase
+
+from src import utils
+
+log = utils.get_pylogger(__name__)
+
+
+@utils.task_wrapper
+def evaluate(cfg: DictConfig) -> Tuple[dict, dict]:
+    """Evaluates given checkpoint on a datamodule testset.
+
+    This method is wrapped in optional @task_wrapper decorator which applies extra utilities
+    before and after the call.
+
+    Args:
+        cfg (DictConfig): Configuration composed by Hydra.
+
+    Returns:
+        Tuple[dict, dict]: Dict with metrics and dict with all instantiated objects.
+    """
+
+    assert cfg.ckpt_path
+
+    log.info(f"Instantiating datamodule <{cfg.datamodule._target_}>")
+    datamodule: LightningDataModule = hydra.utils.instantiate(cfg.datamodule)
+
+    log.info(f"Instantiating model <{cfg.model._target_}>")
+    model: LightningModule = hydra.utils.instantiate(cfg.model)
+
+    log.info("Instantiating loggers...")
+    logger: List[LightningLoggerBase] = utils.instantiate_loggers(cfg.get("logger"))
+
+    log.info(f"Instantiating trainer <{cfg.trainer._target_}>")
+    trainer: Trainer = hydra.utils.instantiate(cfg.trainer, logger=logger)
+
+    object_dict = {
+        "cfg": cfg,
+        "datamodule": datamodule,
+        "model": model,
+        "logger": logger,
+        "trainer": trainer,
+    }
+
+    if logger:
+        log.info("Logging hyperparameters!")
+        utils.log_hyperparameters(object_dict)
+
+    log.info("Starting testing!")
+    trainer.test(model=model, datamodule=datamodule, ckpt_path=cfg.ckpt_path)
+
+    # for predictions use trainer.predict(...)
+    # predictions = trainer.predict(model=model, dataloaders=dataloaders, ckpt_path=cfg.ckpt_path)
+
+    metric_dict = trainer.callback_metrics
+
+    return metric_dict, object_dict
 
 
 @hydra.main(version_base="1.2", config_path=root / "configs", config_name="eval.yaml")
 def main(cfg: DictConfig) -> None:
-
-    from src.tasks.eval_task import evaluate
-
     evaluate(cfg)
