+++ src/train.py
@@ -1,32 +1,127 @@
-# this file acts as a robust starting point for launching hydra runs and multiruns
-# can be run from any place
+import pyrootutils
+
+# -------------------------------------------------------------------------------------
+# `pyrootutils.setup_root(...)` is recommended at the top of each entry file!
+# https://github.com/ashleve/pyrootutils
+#
+# searches for ".git" or "pyproject.toml" in present and parent dirs
+# to determine the project root dir
+#
+# adds root dir to the PYTHONPATH (if `pythonpath=True`)
+# so this file can be run from any place without installing project as a package
+# otherwise modules would need to be imported relatively to the current work dir
+#
+# sets PROJECT_ROOT environment variable
+# used in "configs/paths/default.yaml"
+# this makes all paths relative to project root
+#
+# additionally loads environment variables from ".env" file (if `dotenv=True`)
+#
+root = pyrootutils.setup_root(
+    search_from=__file__,
+    indicator=[".git", "pyproject.toml"],
+    pythonpath=True,
+    dotenv=True,
+)
+#
+# you can get away without using `pyrootutils.setup_root(...)` if you:
+# - move this file to the project root dir
+# - always run this file from the project root dir
+# - modify paths in "configs/paths/default.yaml" to not use PROJECT_ROOT
+#
+# -------------------------------------------------------------------------------------
+
+from typing import List, Optional, Tuple
 
 import hydra
-import pyrootutils
+import pytorch_lightning as pl
 from omegaconf import DictConfig
+from pytorch_lightning import Callback, LightningDataModule, LightningModule, Trainer
+from pytorch_lightning.loggers import LightningLoggerBase
 
-# project root setup
-# searches for root indicators in parent dirs, like ".git", "pyproject.toml", etc.
-# sets PROJECT_ROOT environment variable (used in `configs/paths/default.yaml`)
-# loads environment variables from ".env" if exists
-# adds root dir to the PYTHONPATH (so this file can be run from any place)
-# https://github.com/ashleve/pyrootutils
-root = pyrootutils.setup_root(__file__, dotenv=True, pythonpath=True)
+from src import utils
+
+log = utils.get_pylogger(__name__)
+
+
+@utils.task_wrapper
+def train(cfg: DictConfig) -> Tuple[dict, dict]:
+    """Trains the model. Can additionally evaluate on a testset, using best weights obtained during
+    training.
+
+    This method is wrapped in optional @task_wrapper decorator which applies extra utilities
+    before and after the call.
+
+    Args:
+        cfg (DictConfig): Configuration composed by Hydra.
+
+    Returns:
+        Tuple[dict, dict]: Dict with metrics and dict with all instantiated objects.
+    """
+
+    # set seed for random number generators in pytorch, numpy and python.random
+    if cfg.get("seed"):
+        pl.seed_everything(cfg.seed, workers=True)
+
+    log.info(f"Instantiating datamodule <{cfg.datamodule._target_}>")
+    datamodule: LightningDataModule = hydra.utils.instantiate(cfg.datamodule)
+
+    log.info(f"Instantiating model <{cfg.model._target_}>")
+    model: LightningModule = hydra.utils.instantiate(cfg.model)
+
+    log.info("Instantiating callbacks...")
+    callbacks: List[Callback] = utils.instantiate_callbacks(cfg.get("callbacks"))
+
+    log.info("Instantiating loggers...")
+    logger: List[LightningLoggerBase] = utils.instantiate_loggers(cfg.get("logger"))
+
+    log.info(f"Instantiating trainer <{cfg.trainer._target_}>")
+    trainer: Trainer = hydra.utils.instantiate(cfg.trainer, callbacks=callbacks, logger=logger)
+
+    object_dict = {
+        "cfg": cfg,
+        "datamodule": datamodule,
+        "model": model,
+        "callbacks": callbacks,
+        "logger": logger,
+        "trainer": trainer,
+    }
+
+    if logger:
+        log.info("Logging hyperparameters!")
+        utils.log_hyperparameters(object_dict)
+
+    if cfg.get("train"):
+        log.info("Starting training!")
+        trainer.fit(model=model, datamodule=datamodule, ckpt_path=cfg.get("ckpt_path"))
+
+    train_metrics = trainer.callback_metrics
+
+    if cfg.get("test"):
+        log.info("Starting testing!")
+        ckpt_path = trainer.checkpoint_callback.best_model_path
+        if ckpt_path == "":
+            log.warning("Best ckpt not found! Using current weights for testing...")
+            ckpt_path = None
+        trainer.test(model=model, datamodule=datamodule, ckpt_path=ckpt_path)
+        log.info(f"Best ckpt path: {ckpt_path}")
+
+    test_metrics = trainer.callback_metrics
+
+    # merge train and test metrics
+    metric_dict = {**train_metrics, **test_metrics}
+
+    return metric_dict, object_dict
 
 
 @hydra.main(version_base="1.2", config_path=root / "configs", config_name="train.yaml")
-def main(cfg: DictConfig) -> float:
-
-    # imports can be nested inside @hydra.main to optimize tab completion
-    # https://github.com/facebookresearch/hydra/issues/934
-    from src.tasks.train_task import train
-    from src.utils import get_metric_value
+def main(cfg: DictConfig) -> Optional[float]:
 
     # train the model
     metric_dict, _ = train(cfg)
 
     # safely retrieve metric value for hydra-based hyperparameter optimization
-    metric_value = get_metric_value(
+    metric_value = utils.get_metric_value(
         metric_dict=metric_dict, metric_name=cfg.get("optimized_metric")
     )
